* refactor how parameters are passed to workflows * introduce specific workflows for specific purposes to make adoption easier * update to PSR v1.0.0
